requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.2
sqlparse==0.5.5

Django
djangorestframework
celery
redis
django-filter
python-dotenv
